FlaskWebProject1/requirements.txt
Flask~=1.1
mypy-extensions~=0.4.3
mysql-connector-python~=8.0.29
Unidecode~=1.3.4